# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateBaseRuleRequest:
    properties:
      created_by:
        type: integer
      description:
        maxLength: 255
        type: string
      effective_from:
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - effective_from
    - name
    type: object
...
  models.CreateRuleActionRequest:
    properties:
      action_type:
        $ref: '#/definitions/models.RuleFieldType'
      action_value:
        maxLength: 100
        minLength: 1
        type: string
      rule_id:
        type: integer
    required:
    - action_type
    - action_value
    type: object
  models.CreateRuleConditionRequest:
    properties:
      condition_operator:
        $ref: '#/definitions/models.RuleOperator'
      condition_type:
        $ref: '#/definitions/models.RuleFieldType'
      condition_value:
        maxLength: 100
        minLength: 1
        type: string
      rule_id:
        type: integer
    required:
    - condition_operator
    - condition_type
    - condition_value
    type: object
  models.CreateRuleRequest:
    properties:
      actions:
        items:
          $ref: '#/definitions/models.CreateRuleActionRequest'
        minItems: 1
        type: array
      conditions:
        items:
          $ref: '#/definitions/models.CreateRuleConditionRequest'
        minItems: 1
        type: array
      rule:
        $ref: '#/definitions/models.CreateBaseRuleRequest'
    required:
    - actions
    - conditions
    - rule
    type: object
...
  models.RuleActionResponse:
    properties:
      action_type:
        $ref: '#/definitions/models.RuleFieldType'
      action_value:
        type: string
      id:
        type: integer
      rule_id:
        type: integer
    type: object
  models.RuleConditionResponse:
    properties:
      condition_operator:
        $ref: '#/definitions/models.RuleOperator'
      condition_type:
        $ref: '#/definitions/models.RuleFieldType'
      condition_value:
        type: string
      id:
        type: integer
      rule_id:
        type: integer
    type: object
  models.RuleFieldType:
    enum:
    - amount
    - name
    - description
    - category
    type: string
    x-enum-varnames:
    - RuleFieldAmount
    - RuleFieldName
    - RuleFieldDescription
    - RuleFieldCategory
  models.RuleOperator:
    enum:
    - equals
    - contains
    - greater
    - lower
    type: string
    x-enum-varnames:
    - OperatorEquals
    - OperatorContains
    - OperatorGreater
    - OperatorLower
  models.RuleResponse:
    properties:
      created_by:
        type: integer
      description:
        type: string
      effective_from:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.UpdateRuleActionRequest:
    properties:
      action_type:
        $ref: '#/definitions/models.RuleFieldType'
      action_value:
        type: string
    type: object
  models.UpdateRuleConditionRequest:
    properties:
      condition_operator:
        $ref: '#/definitions/models.RuleOperator'
      condition_type:
        $ref: '#/definitions/models.RuleFieldType'
      condition_value:
        type: string
    type: object
  models.UpdateRuleRequest:
    properties:
      description:
        maxLength: 255
        type: string
      effective_from:
        type: string
      name:
        maxLength: 100
        type: string
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: Enter your email as username and password to authenticate
  title: Expense Tracker API
  version: "1.0"
paths:
...
  /rule:
    get:
      description: Get all transaction categorization rules for the authenticated
        user
      produces:
      - application/json
      responses:
        "200":
          description: List of rules
          schema:
            items:
              $ref: '#/definitions/models.RuleResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BasicAuth: []
      summary: List all rules
      tags:
      - rules
    post:
      consumes:
      - application/json
      description: Create a new transaction categorization rule for the authenticated
        user
      parameters:
      - description: Rule data
        in: body
        name: rule
        required: true
        schema:
          $ref: '#/definitions/models.CreateRuleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Rule created successfully
          schema:
            $ref: '#/definitions/models.RuleResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BasicAuth: []
      summary: Create a new rule
      tags:
      - rules
  /rule/{ruleId}:
    delete:
      description: Delete rule by rule ID for the authenticated user
      parameters:
      - description: Rule ID
        in: path
        name: ruleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Rule deleted successfully
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Rule not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BasicAuth: []
      summary: Delete rule
      tags:
      - rules
    get:
      description: Get rule details by rule ID for the authenticated user
      parameters:
      - description: Rule ID
        in: path
        name: ruleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Rule details
          schema:
            $ref: '#/definitions/models.RuleResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Rule not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BasicAuth: []
      summary: Get rule by ID
      tags:
      - rules
    patch:
      consumes:
      - application/json
      description: Update rule details by rule ID for the authenticated user
      parameters:
      - description: Rule ID
        in: path
        name: ruleId
        required: true
        type: integer
      - description: Updated rule data
        in: body
        name: rule
        required: true
        schema:
          $ref: '#/definitions/models.UpdateRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Rule updated successfully
          schema:
            $ref: '#/definitions/models.RuleResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Rule not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BasicAuth: []
      summary: Update rule
      tags:
      - rules
  /rule/{ruleId}/action/{id}:
    patch:
      consumes:
      - application/json
      description: Update a specific action within a rule for the authenticated user
      parameters:
      - description: Rule ID
        in: path
        name: ruleId
        required: true
        type: integer
      - description: Action ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated action data
        in: body
        name: action
        required: true
        schema:
          $ref: '#/definitions/models.UpdateRuleActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Rule action updated successfully
          schema:
            $ref: '#/definitions/models.RuleActionResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Rule or action not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BasicAuth: []
      summary: Update rule action
      tags:
      - rules
  /rule/{ruleId}/condition/{id}:
    patch:
      consumes:
      - application/json
      description: Update a specific condition within a rule for the authenticated
        user
      parameters:
      - description: Rule ID
        in: path
        name: ruleId
        required: true
        type: integer
      - description: Condition ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated condition data
        in: body
        name: condition
        required: true
        schema:
          $ref: '#/definitions/models.UpdateRuleConditionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Rule condition updated successfully
          schema:
            $ref: '#/definitions/models.RuleConditionResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Rule or condition not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BasicAuth: []
      summary: Update rule condition
      tags:
      - rules
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"